💄 style: 카드 컴포넌트 구조 변경 /common/card -> /card
