+++ src/components/card/admin/adminEntryCard/AdminEntryCard.style.ts
