+++ src/components/card/admin/adminMessageCard/AdminMessageCard.style.ts
