+++ src/components/card/admin/adminSessionCard/AdminSessionCard.tsx
@@ -1,14 +1,15 @@
-import * as S from './AdminMessageCard.style';
+import * as S from './AdminSessionCard.style';
 import Profile from '../../profile/Profile';
-import { Tag } from '../../../tag/Tag.style';
+import Btn from '../../../common/button/btn/Btn';
+import { Tag } from '../../../common/tag/Tag';
 
-type AdminMessageCardProps = {
+type AdminSessionCardProps = {
   title: string;
   name: string;
   from: string;
 };
 
-const AdminMessageCard: React.FC<AdminMessageCardProps> = ({
+const AdminSessionCard: React.FC<AdminSessionCardProps> = ({
   title,
   name,
   from,
@@ -26,8 +27,17 @@ const AdminMessageCard: React.FC<AdminMessageCardProps> = ({
         </S.HeaderContainer>
         <Profile name={name} from={from} />
       </S.ContentsContainer>
+
+      <S.BtnContainer>
+        <Btn variant="secondary" state="default">
+          비활성화
+        </Btn>
+        <Btn variant="primary" state="default">
+          기기연결
+        </Btn>
+      </S.BtnContainer>
     </S.CardContainer>
   );
 };
 
-export default AdminMessageCard;
+export default AdminSessionCard;
