+++ src/components/card/user/faceDataCard/FaceDataCard.tsx
@@ -1,5 +1,5 @@
 import { useTheme } from 'styled-components';
-import Icon from '../../../icon/Icon';
+import Icon from '../../../common/icon/Icon';
 import * as S from './FaceDataCard.style';
 
 type FaceDataCardProps = {};
