+++ src/components/card/user/reservationCard/ReservationCard.tsx
@@ -1,5 +1,6 @@
 import * as S from './ReservationCard.style';
-import { Checkbox } from '../../../checkbox/Checkbox';
+import { Checkbox } from '../../../common/checkbox/Checkbox';
+
 type ReservationCardProps = {
   item: {
     id: number;
